Organize the code and removing redundency
